requirements.txt
anthropic>=0.40.0
rich>=13.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-whois>=0.9.0
python-nmap>=0.7.1
pyfiglet>=1.0.2
psycopg[binary]>=3.2.0
